refactor: Add avatarUsuarioInitState and actualizarAvatarUsuario functions to avatar-usuario.service.ts
This commit adds the avatarUsuarioInitState constant and the actualizarAvatarUsuario function to the avatar-usuario.service.ts file. The avatarUsuarioInitState represents the initial state for the avatar of a user, while the actualizarAvatarUsuario function is responsible for updating the avatar of a user by making a PATCH request to the server. These changes enhance the functionality of the service and provide a way to update avatars for user profiles.