refactor: Improve mobile menu functionality and styling
The mobile menu functionality and styling have been refactored to enhance the user experience. The closeMobileMenu function has been replaced with a simpler and more efficient approach using jQuery's fadeIn and fadeOut methods. Additionally, the saly-link click event now directly sets the display property of the mobileMenu to "none", improving the responsiveness of the menu. These changes improve the overall usability and performance of the mobile menu.