refactor: Add ButtonAddFavorite component for adding books to favorites
Add a new component, ButtonAddFavorite, to allow users to add books to their favorites. This component is used in the book page to provide a button for adding the book to the user's favorites list. It enhances the user experience by providing a convenient way to save and access favorite books.